+++ lib/src/multiple/content_dialog_multiple.dart
@@ -36,6 +36,7 @@ class ContentMultiple extends StatefulWidget {
     required this.unselectedInsideBoxTextStyle,
     required this.updateSelectedItems,
     required this.width,
+    required this.minHeight,
   });
 
   final bool addMode;
@@ -69,6 +70,7 @@ class ContentMultiple extends StatefulWidget {
   final TextStyle? unselectedInsideBoxTextStyle;
   final Function(List<ValueItem>) updateSelectedItems;
   final double width;
+  final double minHeight;
 
   @override
   State<ContentMultiple> createState() => _ContentMultipleState();
@@ -148,8 +150,8 @@ class _ContentMultipleState extends State<ContentMultiple> {
                       widget.dialogSearchBarColor ?? Colors.white),
                   overlayColor: MaterialStatePropertyAll(
                       widget.deactivateHoverColor ?? Colors.grey.shade100),
-                  constraints:
-                      BoxConstraints(maxHeight: 50, maxWidth: widget.width),
+                  constraints: BoxConstraints(
+                      minHeight: widget.minHeight, maxWidth: widget.width),
                   surfaceTintColor: MaterialStatePropertyAll(
                       widget.dialogSearchBarColor ?? Colors.white),
                   shape: MaterialStateProperty.all<OutlinedBorder>(