+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.8.1] - 2023-08-27.
+-   Added `outsideIconColor` property to single and multiple search widgets. It allows the user to set a color to the dropdown arrow. 
+-   Added `outsideIconSize` property to single search widget. It allows the user to set a size to the dropdown arrow. 
+
+## [1.8.0] - 2023-08-27.
+-   Added `selectedItem` property to single search widget. It allows the user to set an initial value to the dropdown. 
+
+## [1.7.0] - 2023-08-27.
+-   Added `clearSelection()` function on multiple and single search widgets. It allows the user to reset the dropdown to the initial state.
+-   Improved sample.
+-   Improved readme.
+
+## [1.6.0] - 2023-08-27.
+-   Made `onAddItem` parameter a `required` one.
+-   Mini fix on searchbar Height
+
 ## [1.5.0] - 2023-08-03.
 -   Convert `List<String>` to `List<ValueItem>`, allowing the users to have anything as an object on the dropdown.